refactor(internal): unify sub-packages and streamline architecture
