+++ internal/templates/contents/profile.yaml
@@ -1,5 +1,5 @@
 site:
-  url: "https://victoriacheng15.vercel.app/"
+  url: "https://victoriacheng15.dev/"
   title: Victoria's Tech Hub
   name: Victoria Cheng
   slogan: Backend systems, infrastructure, and the "why" behind them in practice.
